+++ CHANGELOG.md
@@ -1,4 +1,13 @@
-# HEAD
+# Version 0.2.0
+
+- New version released on 05/9/22
+- Now the application is more user friendly, using CommandLineParser to parse argument and commands. See README.
+- New feature ``demo`` now generates a simple demo image, with the possibility to choose the size of the image and the angle for the observer.
+- The old feature is now named ``pfm2png``.
+- Internally added new geometry class, with shapes Sphere and Plane.
+
+
+# Previous fixings
 
 - Fixed bug in issue https://github.com/Dima0012/DMF-ImaGenerator/issues/3